chore: udpated options for better parameters
